extract the validator to its own file
